Fix unsigned warnings.

+++ src/nh99/model43.cpp
@@ -43,7 +43,7 @@ void param_init_vector_vector::allocate(
   int size = max1 - min1 + 1;
   if (size > 0)
   {
-    v = new param_init_vector[size];
+    v = new param_init_vector[static_cast<unsigned int>(size)];
     if (!v)
     {
       cerr << " error trying to allocate memory in "
